fix: tests for utils already work
